summary/requirements.txt
flask
python-etcd